pyproject.toml
[tool.poetry]
name = "interview-fliits"
version = "0.1.0"
description = ""
authors = ["Pere Rumbo <pere@fliits.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
